+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## v1.3.3
+- Add `contentLength` and `contentMd5` fields to responses of `putJson()`,
+  `postJson()`, and `patchJson()`.
+
 ## v1.2.2
 - Add `conditionalRequestGuards()` API for ExpressJS handlers.
 - Add workaround for `PUT` returning 304 from Minio.